Create FederatedValueProperty
Script is from SAIC validation suite. 

This value property is owned by a block that realizes an external block.  It realizes a value property but their types are not identical.  This rule is only applied if the owning blocks have the same architecture stereotype applied (logical/logical or physical/physical)